examples/extras/abstract_factory_parameters.py
import grpc
from senzing import SzError

from senzing_grpc import SzAbstractFactoryGrpc, SzAbstractFactoryParametersGrpc

factory_parameters: SzAbstractFactoryParametersGrpc = {
    "grpc_channel": grpc.insecure_channel("localhost:8261"),
}

try:
    sz_abstract_factory = SzAbstractFactoryGrpc(**factory_parameters)
except SzError as err:
    print(f"\nERROR: {err}\n")
