requirements.txt
grpcio==1.70.0
protobuf==6.30.2
senzing==0.2.6
